<Form action> can take a relative value
Note the action points to 'destroy'. Like <Link to>, <Form action> can take a relative value. Since the form is rendered in contacts..tsx, then a relative action with destroy will submit the form to contacts..destroy when clicked. After the action redirects, Remix calls all the loaders for the data on the page to get the latest values (this is 'revalidation'). useLoaderData returns new values and causes the components to update!