Forms Without Navigation: useFetcher()
Our new <fetcher.Form method='post'> works almost exactly like the <Form>. It calls the action and then all data is revalidated automatically, even your errors will be caught the same way. There is one key difference though, it's not a navigation, so the URL doesn't change and the history stack is unaffected.